Fixed docs, merged read and mutation modules
