+++ google-cloud-bigtable/lib/google/cloud/bigtable/row_range.rb
@@ -29,8 +29,7 @@ module Google
       #   require "google/cloud/bigtable"
       #
       #   bigtable = Google::Cloud::Bigtable.new
-      #   client = bigtable.client("my-instance")
-      #   table = client.table("my-table")
+      #   table = bigtable.table("my-instance", "my-table", skip_lookup: true)
       #
       #   # Range that includes all row keys including "user-001" to "user-005"
       #   table.row_range.from("user-001").to("user-005", inclusive: true)
@@ -71,8 +70,7 @@ module Google
         #   require "google/cloud/bigtable"
         #
         #   bigtable = Google::Cloud::Bigtable.new
-        #   client = bigtable.client("my-instance")
-        #   table = client.table("my-table")
+        #   table = bigtable.table("my-instance", "my-table", skip_lookup: true)
         #
         #   range = table.row_range.from("key-001")
         #
@@ -80,8 +78,7 @@ module Google
         #   require "google/cloud/bigtable"
         #
         #   bigtable = Google::Cloud::Bigtable.new
-        #   client = bigtable.client("my-instance")
-        #   table = client.table("my-table")
+        #   table = bigtable.table("my-instance", "my-table", skip_lookup: true)
         #
         #   range = table.row_range.from("key-001", inclusive: false)
         #
@@ -105,8 +102,7 @@ module Google
         #   require "google/cloud/bigtable"
         #
         #   bigtable = Google::Cloud::Bigtable.new
-        #   client = bigtable.client("my-instance")
-        #   table = client.table("my-table")
+        #   table = bigtable.table("my-instance", "my-table", skip_lookup: true)
         #
         #   range = table.row_range.to("key-001", inclusive: true)
         #
@@ -114,8 +110,7 @@ module Google
         #   require "google/cloud/bigtable"
         #
         #   bigtable = Google::Cloud::Bigtable.new
-        #   client = bigtable.client("my-instance")
-        #   table = client.table("my-table")
+        #   table = bigtable.table("my-instance", "my-table", skip_lookup: true)
         #
         #   range = table.row_range.to("key-001")
         #
@@ -139,8 +134,7 @@ module Google
         #   require "google/cloud/bigtable"
         #
         #   bigtable = Google::Cloud::Bigtable.new
-        #   client = bigtable.client("my-instance")
-        #   table = client.table("my-table")
+        #   table = bigtable.table("my-instance", "my-table", skip_lookup: true)
         #
         #   range = table.row_range.between("key-001", "key-010")
         #
@@ -159,8 +153,7 @@ module Google
         #   require "google/cloud/bigtable"
         #
         #   bigtable = Google::Cloud::Bigtable.new
-        #   client = bigtable.client("my-instance")
-        #   table = client.table("my-table")
+        #   table = bigtable.table("my-instance", "my-table", skip_lookup: true)
         #
         #   range = table.row_range.of("key-001", "key-010")
         #
